+++ tests/testthat/test-plot-missing.R
@@ -2,7 +2,7 @@ library("states")
 library("ggplot2")
 library("dplyr")
 
-context("`plot_missing`")
+context("plot_missing()")
 
 cy <- state_panel(as.Date("1980-06-30"), as.Date("2015-06-30"), by = "year",
                   useGW = TRUE)
@@ -13,23 +13,23 @@ test_that("plot_missing accepts all input options", {
 
   expect_type(
     plot_missing(data = cy, "myvar", space = "gwcode", time = "date",
-                 time_unit = "year", statelist = "none"),
+                 period = "year", statelist = "none"),
     "list")
 
   expect_type(
     plot_missing(data = cy, "myvar", space = "gwcode", time = "date",
-                 time_unit = "year", statelist = "GW"),
+                 period = "year", statelist = "GW"),
     "list")
 
   expect_type(
     plot_missing(data = cy, "myvar", space = "gwcode", time = "date",
-                 time_unit = "year", statelist = "COW"),
+                 period = "year", statelist = "COW"),
     "list")
 
   # not a valid statelist argument
   expect_error(
     plot_missing(data = cy, "myvar", space = "gwcode", time = "date",
-                 time_unit = "year", statelist = "Foo")
+                 period = "year", statelist = "Foo")
     )
 })
 
@@ -39,14 +39,14 @@ test_that("mssng_mat throws errors for missing ID values", {
     cy2 <- cy
     cy2$gwcode[1] <- NA
     mm <- mssng_mat(data = cy2, "myvar", space = "gwcode", time = "date",
-                    time_unit = "year", statelist = "GW")
+                    period = "year", statelist = "GW")
   })
 
   expect_error({
     cy2 <- cy
     cy2$date[1] <- NA
     mm <- mssng_mat(data = cy2, "myvar", space = "gwcode", time = "date",
-                    time_unit = "year", statelist = "GW")
+                    period = "year", statelist = "GW")
   })
 })
 
@@ -61,4 +61,55 @@ test_that("plot_missing works with tibbles", {
     missing_info(as_tibble(cy), "myvar", "gwcode", "date", "year", "GW")
   )
 
+})
+
+
+context("missing_info()")
+
+test_that("missing_info input checks", {
+
+  cy <- state_panel(as.Date("1980-06-30"), as.Date("2015-06-30"), by = "year",
+                    useGW = TRUE)
+  cy$myvar <- rnorm(nrow(cy))
+  cy$myvar[sample(1:nrow(cy), nrow(cy)*.1, replace = FALSE)] <- NA
+
+  expect_error(missing_info(cy, "myvar", "gwcode", "date", "year", "GW"),
+               NA)
+
+  cy2 <- cy
+  cy2$gwcode[1] <- NA
+  expect_error(missing_info(cy2, "myvar", "gwcode", "date", "year", "GW"),
+               "Space identifier")
+
+  cy2 <- cy
+  cy2$date[1] <- NA
+  expect_error(missing_info(cy2, "myvar", "gwcode", "date", "year", "GW"),
+               "Time identifier")
+
+  cy2 <- cy
+  cy2$date <- as.integer(cy2$date)
+  expect_error(missing_info(cy2, "myvar", "gwcode", "date", "year", "GW"),
+               "Time identifier")
+
+  expect_error(missing_info(cy, "myvar", "gwcode", "date", "year", "G&W"),
+               "Valid choices")
+
+})
+
+test_that("auto input recognition works", {
+  cy <- state_panel(as.Date("1980-06-30"), as.Date("2015-06-30"), by = "year",
+                    useGW = TRUE)
+  cy$myvar <- rnorm(nrow(cy))
+  cy$myvar[sample(1:nrow(cy), nrow(cy)*.1, replace = FALSE)] <- NA
+
+  expect_error(missing_info(cy, "myvar", space = "gwcode", time = "date", period = "year", statelist = "GW"), NA)
+  expect_error(missing_info(cy, "myvar", space = NULL, time = "date", period = "year", statelist = "GW"), NA)
+  cy$cowcode <- cy$gwcode
+  expect_warning(missing_info(cy, "myvar", space = NULL, time = "date", period = "year", statelist = "GW"), "Found both")
+  cy$gwcode <- NULL
+  expect_error(missing_info(cy, "myvar", space = NULL, time = "date", period = "year", statelist = "GW"), NA)
+
+
+  expect_error(missing_info(cy, statelist = "GW"), NA)
+  expect_error(plot_missing(cy, statelist = "GW"), NA)
 })